Moving Mixins to avoid circular dependency
